Ongoing work, tests.

+++ src/main/java/stream/DeltaStreams.java
@@ -10,7 +10,7 @@ import java.util.stream.*;
 public class DeltaStreams {
 
     private static final Random RND = new Random();
-    private static final List<Integer> SIZES = IntStream.range(100, 9900).boxed().collect(Collectors.toList());
+    private static final List<Long> SIZES = LongStream.range(100, 9900).boxed().collect(Collectors.toList());
     private static final List<Double> PRICES = DoubleStream.iterate(975, d -> d + 0.5).limit(100).boxed().collect(Collectors.toList());
 
     public static Stream<Delta> getDummyDeltaStream(){
